add socks5

+++ build.sh
@@ -1,14 +1,13 @@
 #!/bin/bash
-BT=Release
-if [[ $# > 0 ]]; then BT=Debug; fi
-# echo $BT
+# set -x
+
 dir="_build_$1"
-[[ $1 == "pi" ]] || export VCPKG_ROOT=/data/vcpkg
+[[ $HOSTNAME == "y700" ]] && export VCPKG_ROOT=/data/vcpkg
 cmake -H. -B$dir \
--DCMAKE_BUILD_TYPE=$BT \
+-DCMAKE_BUILD_TYPE=Release \
 -G Ninja 
 
 cmake --build $dir
 mkdir -p dist
-mv $dir/py dist/
-[ ! -d dist/prefab ] && cp -r prefab dist/
+[[ -f "$dir/py" ]] && mv $dir/py dist/
+# [ ! -d dist/prefab ] && cp -r prefab dist/
